project-int

+++ src/main/java/com/sh/sh/pos/system/payload/dto/ShiftReportDTO.java
@@ -24,7 +24,7 @@ public class ShiftReportDTO {
 	
 	private Double netSale;
 	
-	private Double totalOrder;
+	private int totalOrder;
 	
 	private UserDTO cashier ;
 	private Long cashierId;
